+++ tests/testthat/test-help.r
@@ -77,3 +77,20 @@ test_that("dev_help works with package and function help with the same name", {
   h1 <- dev_help("testHelp")
   expect_identical(shim_question(testHelp::testHelp), h1)
 })
+
+test_that("unknown macros don't trigger warnings (#119)", {
+  load_all("testUnknownMacro")
+
+  expect_no_warning(
+    out <- dev_help("testUnknownMacro")
+  )
+
+  # Because we're testing internal behaviour in `tools`
+  skip_on_cran()
+
+  # We should still be displaying a warning when rendering the documentation
+  local_options(pager = function(...) "")
+  suppress_output(
+    expect_warning(print(out), "unknown macro")
+  )
+})