+++ tests/testthat/testUnknownMacro/R/testUnknownMacro.R
@@ -1,0 +1,4 @@
+#' Topic
+#'
+#' This macro is \unknown{}.
+testUnknownMacro <- function() {}